Abstracted sign up and login functionality to a component to reuse in auth modal and /signup /login pages
